python_templates/aws_lambda_fast_api/pyproject.toml
[project]
name = "aws-lambda-fastapi-template"
version = "0.1.0"
description = "A lightweight template for building FastAPI applications deployed on AWS Lambda"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "setuptools>=76.0.0",
    "uvicorn>=0.34.0",
    "zappa>=0.59.0",
    "mangum>=0.17.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.1",
    "loguru>=0.7.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "B"]
ignore = []

[tool.ruff.isort]
known-first-party = ["core"] 